upgrade spi input cache length and simplify st7735 driver

+++ spi/spi.cpp
@@ -37,7 +37,7 @@ void Spi::init(uint16_t direction, uint16_t mode, uint16_t data_size,
 	SPI_Cmd(_spi, ENABLE);
 }
 
-void Spi::write8(const uint8_t * out, uint8_t length, uint8_t * in) {
+void Spi::write8(const uint8_t * out, uint16_t length, uint8_t * in) {
 	SPI_DataSizeConfig(_spi, SPI_DataSize_8b );
 	while (length--) {
 		while (!SPI_I2S_GetFlagStatus(_spi, SPI_I2S_FLAG_TXE ))
@@ -51,7 +51,7 @@ void Spi::write8(const uint8_t * out, uint8_t length, uint8_t * in) {
 	}
 }
 
-void Spi::write16(const uint16_t * out, uint8_t length, uint16_t * in) {
+void Spi::write16(const uint16_t * out, uint16_t length, uint16_t * in) {
 	SPI_DataSizeConfig(_spi, SPI_DataSize_16b );
 	while (length--) {
 		while (!SPI_I2S_GetFlagStatus(_spi, SPI_I2S_FLAG_TXE ))
